Move lib/client/BaseGraphqlCapsule.js below lib/client/graphql/
