+++ lib/client/BaseGraphqlLauncher.js
@@ -575,7 +575,7 @@ export default class BaseGraphqlLauncher {
  *   rawResponse: Response | null
  *   payload: furo.Payload<*> | null
  *   result: object | null
- *   abortedReason?: import('./BaseGraphqlCapsule').LAUNCH_ABORTED_REASON
+ *   abortedReason?: import('./graphql/BaseGraphqlCapsule.js').LAUNCH_ABORTED_REASON
  * }} CapsuleParams
  */
