+++ lib/client/graphql/BaseGraphqlCapsule.js
