+++ tests/__tests__/lib/client/BaseGraphqlCapsule.js
@@ -4,7 +4,7 @@ import {
 
 import BaseGraphqlCapsule, {
   LAUNCH_ABORTED_REASON,
-} from '~/lib/client/BaseGraphqlCapsule.js'
+} from '~/lib/client/graphql/BaseGraphqlCapsule.js'
 
 import BaseGraphqlPayload from '~/lib/client/BaseGraphqlPayload.js'
